# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: vcdeploy API
  description: |
    REST API for vcdeploy - a modern deployment orchestration platform.
    
    ## Authentication
    
    Most endpoints require authentication via one of:
    - Session cookie (for web UI)
    - Bearer token in `Authorization` header (for API access)
    - API key in `X-API-Key` header
    
    ## Rate Limiting
    
    API requests are rate limited. When exceeded, endpoints return `429 Too Many Requests`.
    
    ## Request IDs
    
    All requests receive an `X-Request-ID` header for correlation and debugging.
  version: 1.0.0
  license:
    name: MIT
    url: https://github.com/BlackOrder/vcdeploy/blob/main/LICENSE
  contact:
    name: BlackOrder
    url: https://github.com/BlackOrder/vcdeploy
servers:
  - url: http://localhost:9000
    description: Local development server
  - url: https://vcdeploy.example.com
    description: Production server (example)
...
paths:
...
  /api/v1/blocked:
    get:
      tags: [Settings]
      summary: List blocked IPs
      description: Returns all currently blocked IP addresses.
      operationId: listBlockedIPs
      responses:
        '200':
          description: List of blocked IPs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BlockedIP'
    post:
      tags: [Settings]
      summary: Block IP address
      description: Manually block an IP address.
      operationId: blockIP
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BlockIPRequest'
      responses:
        '201':
          description: IP blocked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BlockedIP'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/blocked/{ip}:
    parameters:
      - name: ip
        in: path
        required: true
        schema:
          type: string
        description: IP address to manage
    get:
      tags: [Settings]
      summary: Get blocked IP details
      description: Get details of a specific blocked IP.
      operationId: getBlockedIP
      responses:
        '200':
          description: Blocked IP details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BlockedIP'
        '404':
          description: IP not found in block list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      tags: [Settings]
      summary: Unblock IP address
      description: Remove an IP address from the block list.
      operationId: unblockIP
      responses:
        '204':
          description: IP unblocked
        '404':
          description: IP not found in block list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKey:
      type: apiKey
      in: header
      name: X-API-Key

  # H8 FIX: Reusable pagination parameters
...
  schemas:
...
    Error:
      type: object
      description: Standard error response for all API endpoints
      required:
        - error
        - message
      properties:
        error:
          type: boolean
          description: Always true for error responses
          example: true
        message:
          type: string
          description: Human-readable error message
          example: "Invalid request parameters"

    # H8 FIX: Add pagination metadata schema
...
    BlockedIP:
      type: object
      properties:
        id:
          type: integer
        ipAddress:
          type: string
        blockedAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
        reason:
          type: string
        manualBlock:
          type: boolean
    BlockIPRequest:
      type: object
      required: [ipAddress]
      properties:
        ipAddress:
          type: string
        reason:
          type: string
        duration:
          type: string
          description: Duration string (e.g., "1h", "24h", "7d")
...
security:
  - bearerAuth: []
  - apiKey: []